requirements.txt
# Core LLM frameworks
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# Hugging Face model support
transformers>=4.42.0
huggingface-hub>=0.23.0
accelerate>=0.30.0
sentencepiece  # needed for some tokenizers
bitsandbytes  # optional, for quantized models

# OpenAI API support
openai>=1.30.0

# Embeddings & text processing
tiktoken>=0.7.0
faiss-cpu>=1.8.0  # for vector search
numpy>=1.26.0
pandas>=2.2.0

# Environment & utilities
python-dotenv>=1.0.1
requests>=2.32.0

# Optional: for async API calls
aiohttp>=3.9.0
langchain_huggingface
sentence-transformers
streamlit

